requirements.txt
fastapi
uvicorn
psutil
wmi
gputil
python-multipart
python-socketio